+++ pool/web/view/page/page_admin_mailings.ml
@@ -247,14 +247,13 @@ let form
       CCOption.map_or
         ~default:false
         (fun dist ->
-          CCList.mem_assoc
-            ~eq:Mailing.Distribution.equal_sortable_field
-            field
-            dist)
+          dist |> find_dist |> CCList.mem_assoc ~eq:equal_sortable_field field)
         distribution
     in
     let distribution_fncs =
-      {js|
+      let open Pool_common.Message.Field in
+      Format.asprintf
+        {js|
         function removeDistribution(e) {
           e.preventDefault();
           var field = e.currentTarget.dataset.field;
@@ -285,7 +284,19 @@ let form
           e.detail.shouldSwap = false
         }
       })
+
+      const random = document.querySelector('[name="%s"]');
+      random.addEventListener("change", (e) => {
+        const inputs = document.querySelectorAll('[name="%s"], [name="%s"], [name="%s"]');
+        inputs.forEach((elm) => {
+          elm.disabled = e.currentTarget.checked
+        })
+      })
     |js}
+        (show RandomOrder)
+        (show DistributionField)
+        (show SortOrder)
+        (array_key Distribution)
     in
     let field_select =
       let default_option =
@@ -336,6 +347,10 @@ let form
               |> CCString.capitalize_ascii)
           ]
       ; p [ txt Pool_common.(Utils.hint_to_string language I18n.Distribution) ]
+      ; checkbox_element
+          ?value:(distribution |> CCOption.map Mailing.Distribution.is_random)
+          language
+          Pool_common.Message.Field.RandomOrder
       ; div
           ~a:
             [ a_class
@@ -345,6 +360,7 @@ let form
                 ; "vertical"
                 ; "flexrow"
                 ; "flex-gap"
+                ; "gap"
                 ]
             ]
           [ div
@@ -371,11 +387,9 @@ let form
               ]
           ]
       ; div
-          [ CCOption.map_or
-              ~default:[]
-              (fun distribution ->
-                CCList.map (distribution_form_field language) distribution)
-              distribution
+          [ Mailing.Distribution.(
+              distribution |> CCOption.map_or ~default:[] find_dist)
+            |> CCList.map (distribution_form_field language)
             |> Component.Sortable.create
                  ~classnames:[ "flexcolumn" ]
                  ~attributes:[ a_id "distribution-list" ]